Refactor semver lib
