+++ disperser/dataapi/queried_operators_handlers.go
@@ -5,16 +5,12 @@ import (
 	"fmt"
 	"net"
 	"sort"
-	"strings"
-	"sync"
 	"time"
 
-	"github.com/Layr-Labs/eigenda/api/grpc/node"
 	"github.com/Layr-Labs/eigenda/core"
+	"github.com/Layr-Labs/eigenda/disperser/common/semver"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 	"github.com/gammazero/workerpool"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 )
 
 type OperatorOnlineStatus struct {
@@ -211,121 +207,16 @@ func (s *server) getOperatorInfo(ctx context.Context, operatorId string) (*core.
 }
 
 func (s *server) scanOperatorsHostInfo(ctx context.Context, logger logging.Logger) (*SemverReportResponse, error) {
-	registrations, err := s.subgraphClient.QueryOperatorsWithLimit(context.Background(), 10000)
+	semvers, err := semver.ScanOperatorsHostInfo(ctx, s.subgraphClient, logger)
 	if err != nil {
-		return nil, fmt.Errorf("failed to fetch indexed registered operator state - %s", err)
+		return nil, err
 	}
-	deregistrations, err := s.subgraphClient.QueryOperatorDeregistrations(context.Background(), 10000)
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch indexed deregistered operator state - %s", err)
-	}
+	s.metrics.UpdateSemverCounts(semvers)
 
-	operators := make(map[string]int)
-
-	// Add registrations
-	for _, registration := range registrations {
-		logger.Info("Operator", "operatorId", string(registration.OperatorId), "info", registration)
-		operators[string(registration.OperatorId)]++
-	}
-	// Deduct deregistrations
-	for _, deregistration := range deregistrations {
-		operators[string(deregistration.OperatorId)]--
-	}
-
-	activeOperators := make([]string, 0)
-	for operatorId, count := range operators {
-		if count > 0 {
-			activeOperators = append(activeOperators, operatorId)
-		}
-	}
-	logger.Info("Active operators found", "count", len(activeOperators))
-
-	var wg sync.WaitGroup
-	var mu sync.Mutex
-	numWorkers := 10
-	operatorChan := make(chan string, len(activeOperators))
-	semvers := make(map[string]int)
-	worker := func() {
-		for operatorId := range operatorChan {
-			operatorInfo, err := s.getOperatorInfo(ctx, operatorId)
-			if err != nil {
-				mu.Lock()
-				semvers["not-found"]++
-				mu.Unlock()
-				continue
-			}
-			operatorSocket := core.OperatorSocket(operatorInfo.Socket)
-			dispersalSocket := operatorSocket.GetDispersalSocket()
-			semver := getSemverInfo(context.Background(), dispersalSocket, operatorId, logger)
-
-			mu.Lock()
-			semvers[semver]++
-			mu.Unlock()
-		}
-		wg.Done()
-	}
-
-	// Launch worker goroutines
-	for i := 0; i < numWorkers; i++ {
-		wg.Add(1)
-		go worker()
-	}
-
-	// Send operator IDs to the channel
-	for _, operatorId := range activeOperators {
-		operatorChan <- operatorId
-	}
-	close(operatorChan)
-
-	// Wait for all workers to finish
-	wg.Wait()
-
-	// Create HostInfoReportResponse instance
 	semverReport := &SemverReportResponse{
 		Semver: semvers,
 	}
-
-	// Publish semver report metrics
-	s.metrics.UpdateSemverCounts(semvers)
-
 	return semverReport, nil
-}
-
-// query operator host info endpoint if available
-func getSemverInfo(ctx context.Context, socket string, operatorId string, logger logging.Logger) string {
-	conn, err := grpc.Dial(socket, grpc.WithTransportCredentials(insecure.NewCredentials()))
-	if err != nil {
-		logger.Error("Failed to dial grpc operator socket", "operatorId", operatorId, "socket", socket, "error", err)
-		return "unreachable"
-	}
-	defer conn.Close()
-	client := node.NewDispersalClient(conn)
-	ctxWithTimeout, cancel := context.WithTimeout(ctx, time.Second * time.Duration(3))
-	defer cancel()
-	reply, err := client.NodeInfo(ctxWithTimeout, &node.NodeInfoRequest{})
-	if err != nil {
-		var semver string
-		if strings.Contains(err.Error(), "Unimplemented") {
-			semver = "<0.8.0"
-		} else if strings.Contains(err.Error(), "DeadlineExceeded") {
-			semver = "timeout"
-		} else if strings.Contains(err.Error(), "Unavailable") {
-			semver = "refused"
-		} else {
-			semver = "error"
-		}
-
-		logger.Warn("NodeInfo", "operatorId", operatorId, "semver", semver, "error", err)
-		return semver
-	}
-
-	// local node source compiles without semver
-	if reply.Semver == "" {
-		reply.Semver = "src-compile"
-	}
-
-	logger.Info("NodeInfo", "operatorId", operatorId, "socker", socket, "semver", reply.Semver, "os", reply.Os, "arch", reply.Arch, "numCpu", reply.NumCpu, "memBytes", reply.MemBytes)
-	return reply.Semver
 }
 
 // method to check if operator is online via socket dial
