+++ disperser/dataapi/subgraph/api.go
@@ -27,7 +27,7 @@ type (
 		QueryOperatorInfoByOperatorIdAtBlockNumber(ctx context.Context, operatorId string, blockNumber uint32) (*IndexedOperatorInfo, error)
 		QueryOperatorAddedToQuorum(ctx context.Context, startBlock, endBlock uint32) ([]*OperatorQuorum, error)
 		QueryOperatorRemovedFromQuorum(ctx context.Context, startBlock, endBlock uint32) ([]*OperatorQuorum, error)
-		QueryOperatorDeregistrations(ctx context.Context, first int) ([]*Operator, error)
+		QueryOperatorsDeregistered(ctx context.Context, first int) ([]*Operator, error)
 	}
 
 	api struct {
@@ -111,7 +111,7 @@ func (a *api) QueryOperators(ctx context.Context, first int) ([]*Operator, error
 	return result.OperatorRegistereds, nil
 }
 
-func (a *api) QueryOperatorDeregistrations(ctx context.Context, first int) ([]*Operator, error) {
+func (a *api) QueryOperatorsDeregistered(ctx context.Context, first int) ([]*Operator, error) {
 	variables := map[string]any{
 		"first": graphql.Int(first),
 	}
@@ -121,7 +121,7 @@ func (a *api) QueryOperatorDeregistrations(ctx context.Context, first int) ([]*O
 		return nil, err
 	}
 
-	return result.OperatorRegistereds, nil
+	return result.OperatorDeregistereds, nil
 }
 
 func (a *api) QueryBatchNonSigningInfo(ctx context.Context, startTime, endTime int64) ([]*BatchNonSigningInfo, error) {
