+++ disperser/dataapi/subgraph/queries.go
@@ -85,7 +85,7 @@ type (
 		OperatorRegistereds []*Operator `graphql:"operatorRegistereds(first: $first)"`
 	}
 	queryOperatorDeregistereds struct {
-		OperatorRegistereds []*Operator `graphql:"operatorDeregistereds(first: $first)"`
+		OperatorDeregistereds []*Operator `graphql:"operatorDeregistereds(first: $first)"`
 	}
 	queryBatchNonSigningOperatorIdsInInterval struct {
 		BatchNonSigningOperatorIds []*BatchNonSigningOperatorIds `graphql:"batches(first: $first, skip: $skip, where: {blockTimestamp_gt: $blockTimestamp_gt})"`
